rename isValidLocale to hasLocale (shorter, more expressive what the fn does)
