+++ docs/src/pages/docs/environments/error-files.mdx
@@ -80,12 +80,12 @@ export default function RootLayout({children}) {
 For the 404 page to render, we need to call the `notFound` function in the root layout when we detect an incoming `locale` param that isn't a valid locale.
 
 ```tsx filename="app/[locale]/layout.tsx"
-import {isValidLocale} from 'next-intl';
+import {hasLocale} from 'next-intl';
 import {notFound} from 'next/navigation';
 import {routing} from '@/i18n/routing';
 
 export default function LocaleLayout({children, params: {locale}}) {
-  if (!isValidLocale(routing.locales, locale)) {
+  if (!hasLocale(routing.locales, locale)) {
     notFound();
   }
