+++ examples/example-app-router-mixed-routing/src/db.ts
@@ -1,5 +1,5 @@
 import {cookies} from 'next/headers';
-import {Locale, isValidLocale} from 'next-intl';
+import {Locale, hasLocale} from 'next-intl';
 import {defaultLocale, locales} from './config';
 
 // This cookie name is used by `next-intl` on the public pages too. By
@@ -11,7 +11,7 @@ const COOKIE_NAME = 'NEXT_LOCALE';
 
 export async function getUserLocale(): Promise<Locale> {
   const candidate = cookies().get(COOKIE_NAME)?.value;
-  return isValidLocale(locales, candidate) ? candidate : defaultLocale;
+  return hasLocale(locales, candidate) ? candidate : defaultLocale;
 }
 
 export async function setUserLocale(locale: string) {
