+++ packages/use-intl/src/core/index.tsx
@@ -19,4 +19,4 @@ export type {default as NestedValueOf} from './utils/NestedValueOf.tsx';
 export {createIntlFormatters as _createIntlFormatters} from './formatters.tsx';
 export {createCache as _createCache} from './formatters.tsx';
 export type {default as AppConfig, Locale, Messages} from './AppConfig.tsx';
-export {default as isValidLocale} from './isValidLocale.tsx';
+export {default as hasLocale} from './hasLocale.tsx';
